demo-spring/demo-spring-mvc/src/main/java/com/ltchen/demo/spring/mvc/controller/SpitterController.java
package com.ltchen.demo.spring.mvc.controller;

import javax.validation.Valid;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.validation.Errors;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.ltchen.demo.spring.mvc.bean.Spitter;
import com.ltchen.demo.spring.mvc.bean.SpitterRepository;

@Controller
@RequestMapping("/spitter")
public class SpitterController {
	
	private SpitterRepository spitterRepository;
	
	@Autowired
	public SpitterController(SpitterRepository spitterRepository) {
		this.spitterRepository = spitterRepository;
	}

	@RequestMapping(value="/register",method=RequestMethod.GET)
	public String register(){
		return "register";
	}
	
	@RequestMapping(value="/register",method=RequestMethod.POST)
	public String register(@Valid Spitter spitter,Errors errors){
		//校验输入参数,如果校验错误则重新返回校验表单
		if(errors.hasErrors()){
			return "register";
		}
		//保存Spitter
		spitterRepository.save(spitter);
		//重定向到基本信息页,防止用户重复提交
		return "redirect:/spitter/"+spitter.getUsername();
	}
	
	@RequestMapping(value="/{username}",method=RequestMethod.GET)
	public String profile(@PathVariable String username,Model model){
		Spitter spitter = spitterRepository.findByUsername(username);
		model.addAttribute(spitter);
		return "profile";
	}
}
